Internals refactor types => proto
